requirements.txt
Django==4.2.1
django-environ==0.10.0
psycopg2==2.9.6
Pillow==9.5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
django-filter==23.2
Faker==18.7.0
factory-boy==3.2.1
facebook-sdk==3.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
dj-rest-auth==4.0.0
django-allauth==0.54.0